// src/api/schemas/putReplacementsByIdBodyOne.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */
import type { PutReplacementsByIdBodyOneMatchingType } from './putReplacementsByIdBodyOneMatchingType';

export type PutReplacementsByIdBodyOne = {
  novelId: string;
  /** @minLength 1 */
  from: string;
  /** @minLength 1 */
  to: string;
  matchingType: PutReplacementsByIdBodyOneMatchingType;
};